+++ src/natpmp.c
@@ -216,7 +216,7 @@ int natpmp_impl_process_mcast_response(pcp_impl_t *impl, const char *buffer, int
 
 int natpmp_impl_check_epoch_time(pcp_impl_t *impl, uint32_t curr_server_time) {
 	PLUM_LOG_VERBOSE("NAT-PMP server epoch time is %u", (unsigned int)curr_server_time);
-	uint32_t curr_client_time = current_timestamp() / 1000; // seconds
+	uint32_t curr_client_time = (uint32_t)(current_timestamp() / 1000); // seconds
 
 	// RFC 6886: Every packet sent by the NAT gateway includes a Seconds Since Start of Epoch
 	// (SSSoE) field. [...] Whenever a client receives any packet from the NAT gateway, either
